refactor: 예외 처리기를 GlobalExceptionHandler로 통합
