+++ src/test/java/roomescape/reservation/repository/dao/ReservationDaoTest.java
@@ -162,6 +162,4 @@ class ReservationDaoTest {
         ReservationEntity entity = reservationDao.findById(savedId).orElseThrow();
         Assertions.assertThat(entity.isCancelled()).isTrue();
     }
-
-
 }
